% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{magic_enum_example}
\alias{magic_enum_example}
\title{Examples of \sQuote{magic_enum} use via \pkg{RcppMagicEnum}}
\usage{
magic_enum_example()
}
\description{
This example, directly almost verbatim from the upstream \code{README.md}
shows several simple usage patterns for enumerations enabled by the package.
}
\examples{
magic_enum_example()
}
