\name{RcppMagicEnum-package}
\alias{RcppMagicEnum-package}
\alias{RcppMagicEnum}
\docType{package}
\title{\packageTitle{RcppMagicEnum}}
\description{\packageDescription{RcppMagicEnum}}
\section{Package Content}{\packageIndices{RcppMagicEnum}}
\author{\packageAuthor{RcppMagicEnum}}
\section{Maintainer}{\packageMaintainer{RcppMagicEnum}}
\keyword{package}
